Break out to `reprioritizeAuthOptions` function to make testing easier
