+++ runtime/auth/http-auth-aws/common/src/aws/smithy/kotlin/runtime/http/auth/EndpointAuth.kt
@@ -38,21 +38,3 @@ public fun mergeAuthOptions(modeled: List<AuthOption>, endpointOptions: List<Aut
 
     return merged
 }
-
-/**
- * Re-prioritize a resolved list of auth options based on a user's preference list
- */
-@InternalApi
-public fun reprioritizeAuthOptions(authSchemePreference: List<AuthSchemeId>, authOptions: List<AuthOption>): List<AuthOption> {
-    // add preferred candidates first
-    val preferredAuthOptions = authSchemePreference.mapNotNull { preferredSchemeId ->
-        val preferredSchemeName = preferredSchemeId.id.substringAfter('#')
-        authOptions.singleOrNull {
-            it.schemeId.id.substringAfter('#') == preferredSchemeName
-        }
-    }
-
-    val nonPreferredAuthOptions = authOptions.filterNot { it in preferredAuthOptions }
-
-    return preferredAuthOptions + nonPreferredAuthOptions
-}
