+++ runtime/auth/http-auth-aws/common/test/aws/smithy/kotlin/runtime/http/auth/EndpointAuthTest.kt
@@ -55,75 +55,4 @@ class EndpointAuthTest {
             assertEquals(expectedOption.attributes, actualOption.attributes)
         }
     }
-
-    @Test
-    fun testReprioritizeAuthOptions() = runTest {
-        val preference = listOf(AuthSchemeId.AwsSigV4Asymmetric)
-        val authOptions = listOf(
-            AuthOption(AuthSchemeId.AwsSigV4),
-            AuthOption(AuthSchemeId.AwsSigV4Asymmetric),
-        )
-
-        val expected = listOf(
-            AuthOption(AuthSchemeId.AwsSigV4Asymmetric),
-            AuthOption(AuthSchemeId.AwsSigV4),
-        )
-
-        assertEquals(expected, reprioritizeAuthOptions(preference, authOptions))
-    }
-
-    @Test
-    fun testReprioritizeAuthOptionsNoOp() = runTest {
-        val preference = listOf(AuthSchemeId.HttpBearer)
-        val authOptions = listOf(
-            AuthOption(AuthSchemeId.AwsSigV4),
-            AuthOption(AuthSchemeId.AwsSigV4Asymmetric),
-        )
-
-        val expected = listOf(
-            AuthOption(AuthSchemeId.AwsSigV4),
-            AuthOption(AuthSchemeId.AwsSigV4Asymmetric),
-        )
-
-        assertEquals(expected, reprioritizeAuthOptions(preference, authOptions))
-    }
-
-    @Test
-    fun testReprioritizeAuthOptionsAllPreferred() = runTest {
-        val preference = listOf(AuthSchemeId.HttpDigest, AuthSchemeId.HttpBasic)
-        val authOptions = listOf(
-            AuthOption(AuthSchemeId.HttpBasic),
-            AuthOption(AuthSchemeId.HttpDigest),
-        )
-
-        val expected = listOf(
-            AuthOption(AuthSchemeId.HttpDigest),
-            AuthOption(AuthSchemeId.HttpBasic),
-        )
-
-        assertEquals(expected, reprioritizeAuthOptions(preference, authOptions))
-    }
-
-    @Test
-    fun testReprioritizeAuthOptionsNoPreference() = runTest {
-        val preference = emptyList<AuthSchemeId>()
-        val authOptions = listOf(
-            AuthOption(AuthSchemeId.AwsSigV4),
-            AuthOption(AuthSchemeId.HttpBearer),
-        )
-
-        val expected = authOptions
-
-        assertEquals(expected, reprioritizeAuthOptions(preference, authOptions))
-    }
-
-    @Test
-    fun testReprioritizeAuthOptionsEmptyAuthOptions() = runTest {
-        val preference = listOf(AuthSchemeId.AwsSigV4)
-        val authOptions = emptyList<AuthOption>()
-
-        val expected = emptyList<AuthOption>()
-
-        assertEquals(expected, reprioritizeAuthOptions(preference, authOptions))
-    }
 }
