+++ runtime/auth/http-auth/api/http-auth.api
@@ -59,3 +59,7 @@ public final class aws/smithy/kotlin/runtime/http/auth/BearerTokenSigner : aws/s
 public abstract interface class aws/smithy/kotlin/runtime/http/auth/CloseableBearerTokenProvider : aws/smithy/kotlin/runtime/http/auth/BearerTokenProvider, java/io/Closeable {
 }
 
+public final class aws/smithy/kotlin/runtime/http/auth/ReprioritizeAuthOptionsKt {
+	public static final fun reprioritizeAuthOptions (Ljava/util/List;Ljava/util/List;)Ljava/util/List;
+}
+
